+++ src/BulletDynamics/ConstraintSolver/btConeTwistConstraint.h
@@ -420,13 +420,13 @@ SIMD_FORCE_INLINE const char*	btConeTwistConstraint::serialize(void* dataBuffer,
 	m_rbAFrame.serialize(cone->m_rbAFrame);
 	m_rbBFrame.serialize(cone->m_rbBFrame);
 	
-	cone->m_swingSpan1 = m_swingSpan1;
-	cone->m_swingSpan2 = m_swingSpan2;
-	cone->m_twistSpan = m_twistSpan;
-	cone->m_limitSoftness = m_limitSoftness;
-	cone->m_biasFactor = m_biasFactor;
-	cone->m_relaxationFactor = m_relaxationFactor;
-	cone->m_damping = m_damping;
+	cone->m_swingSpan1 = (float) m_swingSpan1;
+	cone->m_swingSpan2 = (float) m_swingSpan2;
+	cone->m_twistSpan = (float) m_twistSpan;
+	cone->m_limitSoftness = (float) m_limitSoftness;
+	cone->m_biasFactor = (float) m_biasFactor;
+	cone->m_relaxationFactor = (float) m_relaxationFactor;
+	cone->m_damping = (float) m_damping;
 
 	return btConeTwistConstraintDataName;
 }
