+++ src/BulletDynamics/ConstraintSolver/btGeneric6DofSpring2Constraint.h
@@ -604,19 +604,19 @@ SIMD_FORCE_INLINE const char* btGeneric6DofSpring2Constraint::serialize(void* da
 	int i;
 	for (i=0;i<3;i++)
 	{
-		dof->m_angularLowerLimit.m_floats[i]       = m_angularLimits[i].m_loLimit;
-		dof->m_angularUpperLimit.m_floats[i]       = m_angularLimits[i].m_hiLimit;
-		dof->m_angularBounce.m_floats[i]           = m_angularLimits[i].m_bounce;
-		dof->m_angularStopERP.m_floats[i]          = m_angularLimits[i].m_stopERP;
-		dof->m_angularStopCFM.m_floats[i]          = m_angularLimits[i].m_stopCFM;
-		dof->m_angularMotorERP.m_floats[i]         = m_angularLimits[i].m_motorERP;
-		dof->m_angularMotorCFM.m_floats[i]         = m_angularLimits[i].m_motorCFM;
-		dof->m_angularTargetVelocity.m_floats[i]   = m_angularLimits[i].m_targetVelocity;
-		dof->m_angularMaxMotorForce.m_floats[i]    = m_angularLimits[i].m_maxMotorForce;
-		dof->m_angularServoTarget.m_floats[i]      = m_angularLimits[i].m_servoTarget;
-		dof->m_angularSpringStiffness.m_floats[i]  = m_angularLimits[i].m_springStiffness;
-		dof->m_angularSpringDamping.m_floats[i]    = m_angularLimits[i].m_springDamping;
-		dof->m_angularEquilibriumPoint.m_floats[i] = m_angularLimits[i].m_equilibriumPoint;
+		dof->m_angularLowerLimit.m_floats[i]       = (float) m_angularLimits[i].m_loLimit;
+		dof->m_angularUpperLimit.m_floats[i]       = (float) m_angularLimits[i].m_hiLimit;
+		dof->m_angularBounce.m_floats[i]           = (float) m_angularLimits[i].m_bounce;
+		dof->m_angularStopERP.m_floats[i]          = (float) m_angularLimits[i].m_stopERP;
+		dof->m_angularStopCFM.m_floats[i]          = (float) m_angularLimits[i].m_stopCFM;
+		dof->m_angularMotorERP.m_floats[i]         = (float) m_angularLimits[i].m_motorERP;
+		dof->m_angularMotorCFM.m_floats[i]         = (float) m_angularLimits[i].m_motorCFM;
+		dof->m_angularTargetVelocity.m_floats[i]   = (float) m_angularLimits[i].m_targetVelocity;
+		dof->m_angularMaxMotorForce.m_floats[i]    = (float) m_angularLimits[i].m_maxMotorForce;
+		dof->m_angularServoTarget.m_floats[i]      = (float) m_angularLimits[i].m_servoTarget;
+		dof->m_angularSpringStiffness.m_floats[i]  = (float) m_angularLimits[i].m_springStiffness;
+		dof->m_angularSpringDamping.m_floats[i]    = (float) m_angularLimits[i].m_springDamping;
+		dof->m_angularEquilibriumPoint.m_floats[i] = (float) m_angularLimits[i].m_equilibriumPoint;
 	}
 	dof->m_angularLowerLimit.m_floats[3]       = 0;
 	dof->m_angularUpperLimit.m_floats[3]       = 0;
