+++ src/LinearMath/btVector3.h
@@ -1351,7 +1351,7 @@ SIMD_FORCE_INLINE	void	btVector3::serialize(struct	btVector3Data& dataOut) const
 {
 	///could also do a memcpy, check if it is worth it
 	for (int i=0;i<4;i++)
-		dataOut.m_floats[i] = m_floats[i];
+		dataOut.m_floats[i] = (float) m_floats[i];
 }
 
 SIMD_FORCE_INLINE void	btVector3::deSerialize(const struct	btVector3Data& dataIn)
